move finra and salesforce searches into core/
